requirements.txt
gym_unity==0.17.0
mlagents==0.17.0

stable_baselines==2.10.0
torch==1.5.0
tensorboardX==2.0
absl-py==0.9.0
alembic==1.4.2
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
cliff==3.3.0
cloudpickle==1.4.1
cmaes==0.5.0
cmd2==1.1.0
colorama==0.4.3
colorlog==4.1.0
decorator==4.4.2
gast==0.3.3
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
joblib==0.15.1
jupyter-client==6.1.3
jupyter-core==4.6.3
keras==2.3.1
keras-applications==1.0.8
keras-base==2.3.1
Keras-Preprocessing==1.1.2
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
neptune-client==0.4.118
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
optuna==1.5.0
parso==0.7.0
pbr==5.4.5
pickleshare==0.7.5
Pillow==7.1.2
prettytable==0.7.2
prompt-toolkit==3.0.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyperclip==1.8.0
pyreadline==2.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.3.17
stevedore==2.0.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==1.0.1
wrapt==1.12.1
